+++ CHANGELOG.md
@@ -7,14 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+* <code>JsonDeserializable</code> interface - marks object as json deserializable, 
+  so it's possible create instance of object from plain json
+* <code>JsonConvertible</code> interface - merge JsonSerializable and JsonDeserializable together. 
 
 ### Changed
+* JSONAPI\Utils\LinksImpl => JSONAPI\LinksTrait
+* JSONAPI\Utils\MetaImpl => JSONAPI\MetaTrait
 
 ### Deprecated 
 
 ### Fixed
 
 ### Removed 
+* JSONAPI\Utils
 
 ### Security
 
@@ -223,7 +229,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 
 ---
-[Unreleased]: https://gitlab.com/bednic/json-api/compare/2.3.3...2.x
+[Unreleased]: https://gitlab.com/bednic/json-api/compare/2.3.4...2.x
+[2.3.4]: https://gitlab.com/bednic/json-api/compare/2.3.3...2.3.4
 [2.3.3]: https://gitlab.com/bednic/json-api/compare/2.3.2...2.3.3
 [2.3.2]: https://gitlab.com/bednic/json-api/compare/2.3.1...2.3.2
 [2.3.1]: https://gitlab.com/bednic/json-api/compare/2.3.0...2.3.1